refactor: PromptInjectionLLMGuardrail
