+++ weave/scorers/__init__.py
@@ -6,13 +6,15 @@ check_litellm_installation()
 from weave.scorers.classification_scorer import (
     MultiTaskBinaryClassificationF1,
 )
-from weave.scorers.guardrails import PromptInjectionLLMGuardrail
 from weave.scorers.hallucination_scorer import HallucinationFreeScorer
 from weave.scorers.json_scorer import ValidJSONScorer
 from weave.scorers.moderation_scorer import (
     OpenAIModerationScorer,
     WeaveBiasScorerV1,
     WeaveToxicityScorerV1,
+)
+from weave.scorers.prompt_injection_guardrail import (
+    PromptInjectionLLMGuardrail,
 )
 from weave.scorers.pydantic_scorer import PydanticScorer
 from weave.scorers.ragas_scorer import (
