--- weave/scorers/guardrails/__init__.py
@@ -1,5 +1,0 @@
-from weave.scorers.guardrails.prompt_injection_guardrail import (
-    PromptInjectionLLMGuardrail,
-)
-
-__all__ = ["PromptInjectionLLMGuardrail"]
