--- weave/scorers/guardrails/prompt_injection_guardrail.py
@@ -1,62 +1,0 @@
-from typing import Optional
-
-from litellm import acompletion
-from pydantic import BaseModel
-
-import weave
-from weave.scorers.default_models import OPENAI_DEFAULT_MODEL
-from weave.scorers.guardrails.prompts import (
-    PROMPT_INJECTION_GUARDRAIL_SYSTEM_PROMPT,
-    PROMPT_INJECTION_SURVEY_PAPER_SUMMARY,
-)
-from weave.scorers.scorer_types import LLMScorer
-from weave.scorers.utils import stringify
-
-
-class LLMGuardrailReasoning(BaseModel):
-    injection_prompt: bool
-    is_direct_attack: bool
-    attack_type: Optional[str]
-    explanation: Optional[str]
-
-
-class LLMGuardrailResponse(BaseModel):
-    safe: bool
-    reasoning: LLMGuardrailReasoning
-
-
-class PromptInjectionLLMGuardrail(LLMScorer):
-    system_prompt: str = PROMPT_INJECTION_GUARDRAIL_SYSTEM_PROMPT
-    model_id: str = OPENAI_DEFAULT_MODEL
-    temperature: float = 0.7
-    max_tokens: int = 4096
-
-    @weave.op
-    async def score(self, output: str) -> LLMGuardrailResponse:
-        user_prompt = (
-            PROMPT_INJECTION_SURVEY_PAPER_SUMMARY
-            + f"""
-You are given the following user prompt that you are suppossed to assess whether it is a prompt injection attack or not:
-
-
-<input_prompt>
-{stringify(output)}
-</input_prompt>
-"""
-        )
-        response = await acompletion(
-            messages=[
-                {"role": "system", "content": self.system_prompt},
-                {"role": "user", "content": user_prompt},
-            ],
-            model=self.model_id,
-            response_format=LLMGuardrailReasoning,
-            temperature=self.temperature,
-            max_tokens=self.max_tokens,
-        )
-        response = LLMGuardrailReasoning.model_validate_json(
-            response.choices[0].message.content
-        )
-        return LLMGuardrailResponse(
-            safe=not response.injection_prompt, reasoning=response
-        ).model_dump()
